feat: more improvements to CharPoolView
Added the start of a more detailed appearance layout for a character.
Added AppearanceProperty with all the possible (known) entries in an appearance struct in a character.
Added Gender, Personality, and Race as enums for the associated values in AppearanceProperty. These enums implement StaticEnum, which has a helper function to find the enum associated with a given integer value.
Added ControlsFX as a dependency for its SegmentedButton.
Improved handling of nickname display.
Added specific handling of appearance values in Character. Added internal map to store those values.